+++ src/collective/documentgenerator/browser/generation_view.py
@@ -30,10 +30,15 @@ class DocumentGenerationView(BrowserView):
         self.context = context
         self.request = request
 
-    def __call__(self):
-        pod_template = self.get_pod_template()
-        output_format = self.get_generation_format()
+    def __call__(self, template_uid=None, output_format=None):
+        pod_template, output_format = self._get_base_args(template_uid, output_format)
         return self.generate_and_download_doc(pod_template, output_format)
+
+    def _get_base_args(self, template_uid, output_format):
+        template_uid = template_uid or self.get_pod_template_uid()
+        pod_template = self.get_pod_template(template_uid)
+        output_format = output_format or self.get_generation_format()
+        return pod_template, output_format
 
     def generate_and_download_doc(self, pod_template, output_format):
         """
@@ -98,12 +103,11 @@ class DocumentGenerationView(BrowserView):
 
         return document_path
 
-    def get_pod_template(self):
+    def get_pod_template(self, template_uid):
         """
         Return the default PODTemplate that will be used when calling
         this view.
         """
-        template_uid = self.get_pod_template_uid()
         catalog = api.portal.get_tool('portal_catalog')
 
         template_brains = catalog.unrestrictedSearchResults(
@@ -235,9 +239,8 @@ class PersistentDocumentGenerationView(DocumentGenerationView):
     Persistent document generation view.
     """
 
-    def __call__(self):
-        pod_template = self.get_pod_template()
-        output_format = self.get_generation_format()
+    def __call__(self, template_uid, output_format):
+        pod_template, output_format = self._get_base_args(template_uid, output_format)
         persisted_doc = self.generate_persistent_doc(pod_template, output_format)
         self.redirects(persisted_doc)