+++ docs/CHANGES.rst
@@ -11,6 +11,7 @@ Changelog
   toLocalizedDate and adapted AT and DX implementations.
   [gbastien]
 
+
 0.2 (2015-09-22)
 ----------------
 - Renamed field `pod_portal_type` to `pod_portal_types` as this field