+++ docs/CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.3 (unreleased)
+0.3 (2015-09-30)
 ----------------
 
 - Extend the base helper view to do @@plone, @@plone_portal_state view available